+++ app/src/main/java/dev/vinicius/busycardapp/presentation/card_creation/section/OptionsSection.kt
@@ -247,7 +247,7 @@ fun ImageFieldMenu(
             onClick = {
                 onChangeImage(
                     CardCreationEvent.FieldEvent.OnImageFieldChange(
-                        size = size.toFloat(),
+                        size = size,
                         uri = imageUri.value,
                     )
                 )
